making the query function reusable
