+++ bin/run
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
-require('..').run()
+require('../src/index').run()
 .catch(require('@oclif/errors/handle'))
