+++ package.json
@@ -39,7 +39,7 @@
     "oclif"
   ],
   "license": "MIT",
-  "main": "src/index.js",
+  "main": "src/query.js",
   "oclif": {
     "bin": "gq"
   },
